fix: MSVC compiler warnings and missing includes
- Add MSVC warning suppressions (/wd4101, /wd4244, /wd4267)
- Add #include <algorithm> for std::clamp and std::all_of
- Remove unused exception variable 'e' in catch blocks

+++ src/utils/progress.cpp
@@ -55,7 +55,7 @@ ProgressBar::~ProgressBar() {
 void ProgressBar::set_progress(size_t progress) {
     current_progress_ = progress;
     if (bar_) {
-        bar_->set_progress(static_cast<float>(progress));
+        bar_->set_progress(progress);
     }
 }
